Refactor: MeetRoomService 에서 RedisRepository 패턴으로 분리
