+++ src/main/java/capstone/relation/meeting/service/MeetRoomService.java
@@ -1,13 +1,9 @@
 package capstone.relation.meeting.service;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-import org.springframework.data.redis.core.HashOperations;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
@@ -24,44 +20,32 @@ import capstone.relation.meeting.dto.response.JoinResponseDto;
 import capstone.relation.meeting.dto.response.MeetingRoomDto;
 import capstone.relation.meeting.dto.response.MeetingRoomListDto;
 import capstone.relation.meeting.repository.MeetRoomRepository;
+import capstone.relation.meeting.repository.RedisRepository;
 import capstone.relation.user.UserService;
 import capstone.relation.user.dto.RoomInfoDto;
 import capstone.relation.user.dto.UserInfoDto;
 import capstone.relation.websocket.SocketRegistry;
 import capstone.relation.workspace.WorkSpace;
 import capstone.relation.workspace.repository.WorkSpaceRepository;
-import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class MeetRoomService {
-	private static final String WORK_KEY = "WORKSPACE_ROOM_PARTICIPANTS";
-	private static final String USER_KEY = "USER_ROOM_MAPPING";
 	private final UserService userService;
 	private final WorkSpaceRepository workSpaceRepository;
 	private final MeetRoomRepository meetRoomRepository;
+	private final RedisRepository redisRepository;
 	private final SocketRegistry socketRegistry;
-	private final RedisTemplate<String, Object> redisTemplate;
 	private final SimpMessagingTemplate simpMessagingTemplate;
-	private HashOperations<String, String, HashMap<String, Set<String>>> workspaceRoomParticipants;
-	//workspaceId, roomId, userIds
-	private HashOperations<String, String, String> userRoomMapping;
-
-	@PostConstruct
-	protected void init() {
-		workspaceRoomParticipants = redisTemplate.opsForHash();
-		userRoomMapping = redisTemplate.opsForHash();
-	}
 
 	@Transactional(readOnly = false)
 	public JoinResponseDto createAndJoinRoom(CreateRoomDto createRoomDto) {
 		Long userId = SecurityUtil.getCurrentUserId();
 		String roomName = createRoomDto.getRoomName();
-		if (roomName == null || roomName.isEmpty()) {
+		if (roomName == null || roomName.isEmpty())
 			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "회의실 이름을 입력해주세요.");
-		}
 		String workSpaceId = userService.getUserWorkSpaceId(userId);
 		try {
 			JoinResponseDto joinResponseDto = createAndJoin(workSpaceId, userId.toString(), roomName);
@@ -75,17 +59,15 @@ public class MeetRoomService {
 	}
 
 	public boolean isUserInRoom(String userId) {
-		return userRoomMapping.get(USER_KEY, userId) != null;
+		return redisRepository.isUserInRoom(userId);
 	}
 
 	public RoomInfoDto getRoomInfo(String workspaceId, Long userId) {
-		if (!isUserInRoom(userId.toString())) {
+		if (!isUserInRoom(userId.toString()))
 			return new RoomInfoDto();
-		}
-		String roomId = userRoomMapping.get(USER_KEY, userId.toString());
-		System.out.println("roomId = " + roomId);
-		System.out.println("userId = " + userId);
-		Set<String> userIds = getRoomMembers(workspaceId, Long.parseLong(roomId));
+
+		String roomId = redisRepository.getUserRoom(userId.toString());
+		Set<String> userIds = redisRepository.getRoomMembers(workspaceId, Long.parseLong(roomId));
 		List<UserInfoDto> userInfoList = userService.getUserInfoList(userIds);
 		return new RoomInfoDto(true, Long.parseLong(roomId),
 			meetRoomRepository.findById(Long.parseLong(roomId)).get().getRoomName(), userInfoList);
@@ -94,9 +76,8 @@ public class MeetRoomService {
 	private JoinResponseDto createAndJoin(String workSpaceId, String userId, String roomName) {
 		WorkSpace workSpace = workSpaceRepository.findById(workSpaceId)
 			.orElseThrow(() -> new IllegalArgumentException("Invalid workspace ID"));
-		if (isUserInRoom(userId)) {
+		if (isUserInRoom(userId))
 			throw new IllegalArgumentException("User is already in the room: " + userId);
-		}
 
 		MeetRoom meetRoom = MeetRoom.builder()
 			.roomName(roomName)
@@ -110,9 +91,8 @@ public class MeetRoomService {
 
 	public JoinResponseDto joinRoom(Long roomId) {
 		Long userId = SecurityUtil.getCurrentUserId();
-		//유저가 이미 회의 방에 있는지 확인
 		String workspaceId = userService.getUserWorkSpaceId(userId);
-		String meetRoom = userRoomMapping.get(USER_KEY, userId.toString());
+		String meetRoom = redisRepository.getUserRoom(userId.toString());
 		if (meetRoom != null)
 			throw new IllegalArgumentException("User is already in the room: " + userId);
 
@@ -124,7 +104,8 @@ public class MeetRoomService {
 	private JoinResponseDto joinWorkspaceRoom(String workSpaceId, String userId, Long roomId) {
 		MeetRoom meetRoom = meetRoomRepository.findById(roomId)
 			.orElseThrow(() -> new IllegalArgumentException("Invalid room ID"));
-		Set<String> userIds = addUserToRoom(workSpaceId, roomId, userId);
+		redisRepository.addUserToRoom(workSpaceId, roomId, userId);
+		Set<String> userIds = redisRepository.getRoomMembers(workSpaceId, roomId);
 		List<UserInfoDto> userInfoList = userService.getUserInfoList(userIds);
 		sendUserList(workSpaceId, roomId);
 		return new JoinResponseDto(roomId, meetRoom.getRoomName(), userInfoList, (long)userIds.size());
@@ -133,63 +114,12 @@ public class MeetRoomService {
 	@Transactional(readOnly = false)
 	public void leaveRoom(Long userId) {
 		String workspaceId = userService.getUserWorkSpaceId(userId);
-		String meetRoom = userRoomMapping.get(USER_KEY, userId.toString());
-		if (meetRoom == null) {
+		String meetRoom = redisRepository.getUserRoom(userId.toString());
+		if (meetRoom == null)
 			return;
-		}
-		removeUserFromRoom(workspaceId, Long.parseLong(meetRoom), userId.toString());
+		redisRepository.removeUserFromRoom(workspaceId, Long.parseLong(meetRoom), userId.toString());
 		sendUserList(workspaceId, Long.parseLong(meetRoom));
 		sendRoomList(workspaceId);
-	}
-
-	private Set<String> addUserToRoom(String workspaceId, Long roomId, String userId) {
-		HashMap<String, Set<String>> roomParticipants = workspaceRoomParticipants.get(WORK_KEY, workspaceId);
-		if (roomParticipants == null) {
-			roomParticipants = new HashMap<>();
-		}
-		Set<String> users = roomParticipants.get(roomId.toString());
-		if (users == null) {
-			users = new HashSet<>();
-		}
-		users.add(userId);
-		roomParticipants.put(roomId.toString(), users);
-		workspaceRoomParticipants.put(WORK_KEY, workspaceId, roomParticipants);
-		userRoomMapping.put(USER_KEY, userId, roomId.toString());
-		return users;
-	}
-
-	private void removeUserFromRoom(String workspaceId, Long roomId, String userId) {
-		userRoomMapping.delete(USER_KEY, userId);
-		HashMap<String, Set<String>> roomParticipants = workspaceRoomParticipants.get(WORK_KEY, workspaceId);
-		if (roomParticipants == null) {
-			return;
-		}
-		Set<String> userIds = roomParticipants.get(roomId.toString());
-		userIds.remove(userId);
-
-		if (userIds.isEmpty()) {
-			MeetRoom meetRoom = meetRoomRepository.findById(roomId).orElse(null);
-			if (meetRoom != null) {
-				meetRoom.setDeleted(true);
-				meetRoomRepository.save(meetRoom);
-			}
-			roomParticipants.remove(roomId.toString());
-		} else {
-			roomParticipants.put(roomId.toString(), userIds);
-		}
-		if (roomParticipants.isEmpty()) {
-			workspaceRoomParticipants.delete(WORK_KEY, workspaceId);
-		} else {
-			workspaceRoomParticipants.put(WORK_KEY, workspaceId, roomParticipants);
-		}
-	}
-
-	public Set<String> getRoomMembers(String workspaceId, Long roomId) {
-		HashMap<String, Set<String>> roomParticipants = workspaceRoomParticipants.get(WORK_KEY, workspaceId.toString());
-		if (roomParticipants == null) {
-			return new HashSet<>();
-		}
-		return roomParticipants.get(roomId.toString());
 	}
 
 	public MeetingRoomListDto sendRoomList(String workSpaceId) {
@@ -200,7 +130,7 @@ public class MeetRoomService {
 			MeetingRoomDto meetingRoomDto = new MeetingRoomDto();
 			meetingRoomDto.setRoomId(roomId);
 			meetingRoomDto.setRoomName(meetRoom.getRoomName());
-			Set<String> roomMembers = getRoomMembers(workSpaceId, roomId);
+			Set<String> roomMembers = redisRepository.getRoomMembers(workSpaceId, roomId);
 			if (roomMembers == null) {
 				return meetingRoomListDto;
 			}
@@ -220,7 +150,7 @@ public class MeetRoomService {
 	}
 
 	private void sendUserList(String workSpaceId, Long roomId) {
-		Set<String> userIds = getRoomMembers(workSpaceId, roomId);
+		Set<String> userIds = redisRepository.getRoomMembers(workSpaceId, roomId);
 		List<UserInfoDto> userInfoList = userService.getUserInfoList(userIds);
 		simpMessagingTemplate.convertAndSend("/topic/meetingRoom/" + roomId + "/users", userInfoList);
 	}
